refactor : menu_app 파일명, 사이즈 변경, header 폴더 위치 변경
