+++ src/components/MenuApp.module.css
@@ -5,7 +5,7 @@
     background-color: #fff;
     padding: 16px 0;
     width: 360px;
-    height: 78px;
+    height: 62px;
     box-shadow: 0 -4px 16px rgba(0, 0, 0, 0.12);
 }
